[svn-r9781] Purpose:     Bug fix
Description:
    Print the thread ID in a little more portable of a fashion, disallowing
negative thread IDs.

Platforms tested:
    Linux 2.4 (heping) w/threadsafe
    Too Minor to require h5committest

+++ src/H5E.c
@@ -472,7 +472,7 @@ H5Eprint(FILE *stream)
 	    fprintf (stream, "thread 0.");
     }
 #elif defined(H5_HAVE_THREADSAFE)
-    fprintf (stream, "thread %d.", (int)pthread_self());
+    fprintf (stream, "thread %lu.", (unsigned long)pthread_self());
 #else
     fprintf (stream, "thread 0.");
 #endif
